Allow the use of Python interpreters that can't import herdsman.loggrok
To allow using an arbitrary Python interpreter (as long as it can run
the contents of loggrok.py), we need to supply the path to loggrok.py
rather than expecting the interpreter to find it. This also means we
need to fiddle with sys.path a bit to make it possible for the user code
to import files from the current directory.

As explained in the comments, I'm unsure why exactly this is, but see
e.g. https://stackoverflow.com/q/49559003/5951320 -- when running a
script via `python -m xyz`, the first entry of sys.path is the empty
string, but when running `python xyz.py`, it's not.